clean up constants
