+++ src/main/preload.ts
@@ -1,14 +1,21 @@
-import {
-  APP_INFO_QUERY,
-  APP_INFO_TYPE,
-  AUDIO_MSG,
-  AUDIO_MSG_TYPE,
-  AUDIO_START,
-  AUDIO_STOP,
-  LOG_MSG,
-  LOG_MSG_TYPE,
-} from '@/shared/constants'
 import { contextBridge, ipcRenderer } from 'electron'
+
+export const APP_INFO_QUERY = 'app_info_query'
+export type APP_INFO_TYPE = { appName: string; appVersion: string }
+export const LOG_MSG = 'log_mmsg'
+export type LOG_MSG_TYPE = {
+  level: 'error' | 'warn' | 'info' | 'debug'
+  message: string
+}
+
+export const AUDIO_START = 'audio_start'
+export const AUDIO_STOP = 'audio_stop'
+export const AUDIO_MSG = 'audio_msg'
+export type AUDIO_MSG_TYPE = {
+  error: string | undefined
+  stderr: string
+  stdout: string
+}
 
 export const preload = {
   PUBLIC_URL: process.env.PUBLIC_URL,
@@ -25,7 +32,6 @@ export const preload = {
   audio: {
     start: () => ipcRenderer.send(AUDIO_START),
     stop: () => ipcRenderer.send(AUDIO_STOP),
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     receive: (cb: (args: AUDIO_MSG_TYPE) => void) => {
       ipcRenderer.on(AUDIO_MSG, (_event, args) => cb(args as AUDIO_MSG_TYPE))
     },
